requirements.txt
biopython==1.81
pandas>=2.1.1
dna_features_viewer==3.1.2
requests==2.31.0
matplotlib==3.7.1